WiP footer goes directly to payment if there is only one payment
