+++ ts/features/pn/components/MessageFooter.tsx
@@ -1,4 +1,4 @@
-import React, { MutableRefObject } from "react";
+import React, { MutableRefObject, useCallback } from "react";
 import { StyleSheet, View } from "react-native";
 import { ButtonSolid, IOStyles } from "@pagopa/io-app-design-system";
 import I18n from "i18n-js";
@@ -7,6 +7,10 @@ import { useIOSelector } from "../../../store/hooks";
 import { UIMessageId } from "../../../store/reducers/entities/messages/types";
 import { paymentsButtonStateSelector } from "../store/reducers/payments";
 import variables from "../../../theme/variables";
+import { initializeAndNavigateToWalleForPayment } from "../utils";
+import { getRptIdStringFromPayment } from "../utils/rptId";
+import { useIOToast } from "../../../components/Toast";
+import { useDispatch } from "react-redux";
 
 const styles = StyleSheet.create({
   container: {
@@ -39,6 +43,19 @@ export const MessageFooter = ({
       maxVisiblePaymentCount
     )
   );
+  const dispatch = useDispatch();
+  const toast = useIOToast();
+  const onFooterPressCallback = useCallback(() => {
+    if (payments?.length === 1) {
+      const firstPayment = payments[0];
+      const paymentId = getRptIdStringFromPayment(firstPayment);
+      initializeAndNavigateToWalleForPayment(paymentId, dispatch, () =>
+        toast.error(I18n.t("genericError"))
+      );
+    } else {
+      presentPaymentsBottomSheetRef.current?.();
+    }
+  }, [dispatch, payments, presentPaymentsBottomSheetRef, toast]);
   if (isCancelled || buttonState === "hidden") {
     return null;
   }
@@ -53,7 +70,7 @@ export const MessageFooter = ({
           loading={isLoading}
           color="primary"
           label={I18n.t("wallet.continue")}
-          onPress={() => presentPaymentsBottomSheetRef.current?.()}
+          onPress={onFooterPressCallback}
           accessibilityLabel={I18n.t("wallet.continue")}
         />
       </View>
