+++ ts/features/pn/utils/index.ts
@@ -1,7 +1,10 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import { identity, pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/lib/Option";
+import * as E from "fp-ts/lib/Either";
 import * as RA from "fp-ts/lib/ReadonlyArray";
+import { RptIdFromString } from "@pagopa/io-pagopa-commons/lib/pagopa";
+import { Dispatch } from "redux";
 import I18n from "../../../i18n";
 import { UIService } from "../../../store/reducers/entities/services/types";
 import { PNMessage } from "../store/types/types";
@@ -11,6 +14,9 @@ import { isServiceDetailNavigationLink } from "../../../utils/internalLink";
 import { GlobalState } from "../../../store/reducers/types";
 import { NotificationRecipient } from "../../../../definitions/pn/NotificationRecipient";
 import { NotificationPaymentInfo } from "../../../../definitions/pn/NotificationPaymentInfo";
+import { paymentInitializeState } from "../../../store/actions/wallet/payment";
+import { navigationRef } from "../../../navigation/NavigationService";
+import ROUTES from "../../../navigation/routes";
 
 export function getNotificationStatusInfo(status: NotificationStatus) {
   return I18n.t(`features.pn.details.timeline.status.${status}`, {
@@ -127,3 +133,20 @@ export const isCancelledFromPNMessagePot = (
     O.chainNullableK(message => message.isCancelled),
     O.getOrElse(() => false)
   );
+
+export const initializeAndNavigateToWalleForPayment = (paymentId: string, dispatch: Dispatch<any>, decodeErrorCallback: (() => void) | undefined, preNavigationCallback: (() => void) | undefined = undefined) => {
+  const eitherRptId = RptIdFromString.decode(paymentId);
+  if (E.isLeft(eitherRptId)) {
+    decodeErrorCallback?.();
+    return;
+  }
+
+  preNavigationCallback?.();
+    
+  dispatch(paymentInitializeState());
+
+  navigationRef.current?.navigate(ROUTES.WALLET_NAVIGATOR, {
+    screen: ROUTES.PAYMENT_TRANSACTION_SUMMARY,
+    params: { rptId: eitherRptId.right }
+  });
+};
